requirements.txt
tensorflow_probability==0.18
tensorflow_addons
vip_hci==1.2.2
pandas
hciplot
notebook
names
